requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
networkx==3.2.1

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0

# Data processing
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
cryptography>=45.0.7,<47

# Optional visualization enhancement
pygraphviz==1.11